Refactored and moved the config row component to main conponents folders to be reused.
